requirements.txt
attrs==21.4.0
certifi==2021.10.8
chardet==4.0.0
click==8.1.2
dataclasses==0.6
deprecation==2.1.0
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
gotrue==0.2.0
greenlet==1.1.2
h11==0.13.0
httpcore==0.12.3
httpx==0.16.1
idna==2.10
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
postgrest-py==0.4.0
py==1.11.0
pyparsing==3.0.8
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.20.0
realtime-py==0.1.3
requests==2.25.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
supabase-py==0.0.2
toml==0.10.2
urllib3==1.26.9
websockets==9.1
Werkzeug==2.1.1
